Test d'integratiin BottomSheetModal pour l'affichage final
