+++ babel.config.js
@@ -1,6 +1,11 @@
 module.exports = function(api) {
   api.cache(true);
   return {
-    presets: ['babel-preset-expo'],
+    presets: [
+      'babel-preset-expo'
+    ],
+    plugins: [
+      'react-native-reanimated/plugin',
+    ],
   };
 };
